requirements.txt
tensorflow-gpu >= 1.13.0